Unit tests

+++ facilities/src/main/java/meet/facilities/util/Constant.java
@@ -4,8 +4,8 @@ public final class Constant {
     public static final int MIN_TEMP = 20;
     public static final int MAX_TEMP = 24;
 	public static final double BEERS_COLD_DAY = 0.75;
-	public static final int BEERS_WARM_DAY = 1;
-	public static final int BEERS_HOT_DAY = 2;
+	public static final double BEERS_WARM_DAY = 1.0;
+	public static final double BEERS_HOT_DAY = 2.0;
 	public static final String FORMAT_DATE = "yyyy-mm-dd";
 
 }
